pyproject.toml
[tool.poetry]
name = "superchat"
version = "0.1.0"
description = ""
authors = ["nakamoo <k00ook17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
flask = {extras = ["async"], version = "^2.3.2"}
pyjsparser = "^2.7.1"
langchain = "^0.0.167"
llama-index = "^0.6.6"
python-dotenv = "^1.0.0"
flask-httpauth = "^4.8.0"
authlib = "^1.2.0"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
bs4 = "^0.0.1"
slimit = "^0.8.1"
scrapy = "^2.9.0"
google-auth = "^2.18.1"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
gunicorn = "^20.1.0"
ipykernel = "^6.23.1"
autoflake = "^2.1.1"
rq = "^1.15.0"
flask-session = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
